requirements.txt
# 這是你的 requirements.txt 檔案內容
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
youtube-transcript-api>=0.6.1
google-genai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0